Extract language server features to separate files
